♻️ 메인 페이지 코드 수정
