+++ src/app/main/index.tsx
@@ -4,94 +4,59 @@
 
 import Link from "next/link";
 import { useRouter } from "next/navigation";
-import { COLOR } from "@/styles/color";
 
-import { Resume } from "@type/resume";
-import { emptyTemplate, basicTemplate } from "@assets/resumeTemplate";
+import type { Resume } from "@type/resume";
+
+import { createResume } from "@/utils/supabase/createResume";
 import Header from "@components/Header";
 import Card, { resumeCard } from "@/components/Card";
-import { createClient } from "@/utils/supabase/client";
-import { cardTitle, grid, resetLinkStyle, secondaryCard } from "./style";
+import {
+  cardTitle,
+  container,
+  grid,
+  resetLinkStyle,
+  secondaryCard,
+} from "./style";
 
-const RESUME_TABLE = "resume";
+const newResumeOptions = [
+  { useTemplate: false, buttonTitle: "처음부터 시작하기" },
+  { useTemplate: true, buttonTitle: "템플릿으로 시작하기" },
+];
 
 export default function MainPage({ resumeList }: { resumeList: Resume[] }) {
   const router = useRouter();
 
-  async function makeNewResume({ useTemplate }: { useTemplate: boolean }) {
-    const supabase = createClient();
-    const {
-      data: { user },
-    } = await supabase.auth.getUser();
-
-    try {
-      const { data, error } = await supabase
-        .from(RESUME_TABLE)
-        .insert({
-          title: useTemplate ? "템플릿으로 시작하기" : "빈 이력서",
-          content: useTemplate ? basicTemplate : emptyTemplate,
-          uid: user.id,
-          modified_at: new Date(),
-          main_color: "#003FC7",
-        })
-        .select()
-        .single();
-
-      if (error) {
-        throw new Error();
-      }
-      return data.id;
-    } catch (error) {
-      alert("새로 이력서를 만들지 못했습니다");
-      return null;
-    }
-  }
-
   return (
-    <div>
+    <div css={container}>
       <Header />
 
-      <div
-        css={{
-          height: "calc(100vh - 4rem)",
-          background: COLOR.LIGHT_GRAY.STANDARD,
-        }}
-      >
-        <div css={grid}>
-          {resumeList?.map((resumeDatum: Resume, resumeDatumIdx: number) => (
-            <Link
-              key={resumeDatumIdx}
-              href={`/build?resumeId=${resumeDatum?.id}`}
-              css={resetLinkStyle}
-            >
-              <Card data={resumeDatum} />
-            </Link>
-          ))}
+      <div css={grid}>
+        {resumeList?.map((resumeDatum: Resume, resumeDatumIdx: number) => (
+          <Link
+            key={resumeDatumIdx}
+            href={`/build?resumeId=${resumeDatum?.id}`}
+            css={resetLinkStyle}
+          >
+            <Card data={resumeDatum} />
+          </Link>
+        ))}
+
+        {newResumeOptions.map((option) => (
           <button
             css={[resumeCard, secondaryCard]}
             onClick={async () => {
-              const id = await makeNewResume({ useTemplate: false });
+              const id = await createResume({
+                useTemplate: option.useTemplate,
+              });
 
               if (id) {
                 router.push(`/build?resumeId=${id}`);
               }
             }}
           >
-            <h2 css={cardTitle}>처음부터 시작하기</h2>
+            <h2 css={cardTitle}>{option.buttonTitle}</h2>
           </button>
-          <button
-            css={[resumeCard, secondaryCard]}
-            onClick={async () => {
-              const id = await makeNewResume({ useTemplate: true });
-
-              if (id) {
-                router.push(`/build?resumeId=${id}`);
-              }
-            }}
-          >
-            <h2 css={cardTitle}>템플릿으로 시작하기</h2>
-          </button>
-        </div>
+        ))}
       </div>
     </div>
   );
