+++ src/styles/index.ts
@@ -1,2 +1,1 @@
-export const HEADER_HEIGHT = "4REM";
 export const ACTION_PANEL_HEIGHT = "3rem";
