+++ src/styles/size.ts
@@ -1,3 +1,4 @@
-export const size = {
-  icon: { small: "1rem" },
+export const SIZE = {
+  HEADER_HEIGHT: "4rem",
+  ICON: { SMALL: "1rem", STANDARD: "2rem" },
 };
